+++ src/lies/scrapers/web.py
@@ -101,7 +101,7 @@ class WebScraper(BaseScraper):
             with urllib.request.urlopen(req) as resp:
                 final_url = resp.geturl()
                 body: bytes = resp.read()
-        except HTTPError, URLError:
+        except (HTTPError, URLError):
             return None
 
         if final_url.rstrip("/") != url.rstrip("/"):
@@ -126,7 +126,7 @@ class WebScraper(BaseScraper):
             with urllib.request.urlopen(req, timeout=30) as resp:
                 final_url = resp.geturl()
                 body: bytes = resp.read()
-        except HTTPError, URLError, TimeoutError, OSError:
+        except (HTTPError, URLError, TimeoutError, OSError):
             return None
 
         if final_url.rstrip("/") != url.rstrip("/"):